ENH: primal and adjoint solvers utilised by adjointOptimisationFoam
will now attempt to read a fall-back divScheme, if the one pertaining to
the exact names of the fields being solved is not found in fvSchemes.

For instance, in multi-point optimisation runs where the fields must be
appended by the solver name (assume it is p1), the user, up to now, had
to provide the corresponding convection scheme, e.g.

div(phip1,Upp1) Gauss upwind;

With the current commit, if the above scheme is not found in fvSchemes,
a fall-back scheme consisting of the typical names that would be used if
useSolverNameForFields is set to false is looked-up as well, e.g.

div(phi,U) Gauss upwind;

The order of precedence in the above mentioned example would be

1. div(phip1,Upp1)
2. div(phi,U)
3. The default scheme in divSchemes

This also holds for the adjoint equations and if mean primal fields
are used for their solution.